+++ gillespy2/solvers/cpp/c_base/model.h
@@ -40,7 +40,7 @@ namespace Gillespy{
     virtual ~IPropensityFunction() {}; 
   };
 
-  
+
   //Represents simulation return data
   struct Simulation{
     Model* model;
@@ -50,15 +50,19 @@ namespace Gillespy{
     double end_time;
     double current_time;
     int random_seed;
-
     unsigned int number_timesteps;
     unsigned int number_trajectories;
-    unsigned int* trajectories_1D;
-    unsigned int*** trajectories;
+    double* trajectories_1D;
+    double*** trajectories;
     IPropensityFunction *propensity_function;
     Simulation(Model* model, unsigned int number_trajectories, unsigned int number_timesteps, double end_time, IPropensityFunction* propensity_function, int random_seed, double current_time, int ode);
     friend std :: ostream& operator<<(std :: ostream& os, const Simulation& simulation);
     void output_results_buffer(std :: ostream& os);
   };
+
+
+
+
+
 }
 #endif
